requirements.txt
streamlit==1.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pdfplumber==0.10.3
pytesseract==0.3.10
PyMuPDF==1.23.8
opencv-python==4.9.0.80
Pillow==10.2.0
langchain==0.1.9
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv==1.0.1
mistralai==0.0.12 